#39726: Refactor of RPC message response waits.
This is attempting to resolve some performance issues found during QA. Rather than relying on a wait duration when awaiting a response from the server, we do a single check for new messages before looping. This also resolves an issue where two process events calls were being made for each wait duration when expecting a response to an RPC command.