#40397 Panel console polish
Adds 2 new buttons to the panel console:

* Select All
* Copy Selected to Clipboard

The old header is also removed and the Close button is moved to the bottom,
right-aligned.

The engine also now forwards the log file path to the js side for display/open
in error screens. Clicking the file path in the error screen will open the
file. On OSX the file is opened in the Console app.

Closes #31.