For SG-2596: Fixes window parenting on Windows when running in a PySide2 environment. (#59)
If the integration is making use of a Python environment that contains PySide2 on Windows, we're required to take a different approach to setting up window parenting for Toolkit apps. The way window ids work in Qt5 is different than in Qt4, and PySide2 does no expose the necessary pieces of the framework to be able to map from a Qt window id to an HWND. As such, we have to take a different approach to getting the HWND of our proxy parent widget than we do when Qt4/PySide is being used.